+++ backend/canisters/user/impl/src/updates/update_btc_balance.rs
@@ -3,7 +3,9 @@ use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use ckbtc_minter_canister::update_balance::{UpdateBalanceError, UtxoStatus};
 use ckbtc_minter_canister::{CKBTC_MINTER_CANISTER_ID, TESTNET_CKBTC_MINTER_CANISTER_ID};
+use event_store_producer::EventBuilder;
 use ledger_utils::format_crypto_amount;
+use serde::Serialize;
 use tracing::error;
 use user_canister::update_btc_balance::{Response::*, *};
 
@@ -43,6 +45,14 @@ Your account has been credited with {formatted} BTC."
                         false,
                         state,
                     );
+                    let user_id_string = state.env.canister_id().to_string();
+                    state.data.event_store_client.push(
+                        EventBuilder::new("btc_deposit", state.env.now())
+                            .with_user(user_id_string.clone(), true)
+                            .with_source(user_id_string, true)
+                            .with_json_payload(&BtcDepositEventPayload { amount: total_minted })
+                            .build(),
+                    )
                 }
                 for error in errors {
                     crate::openchat_bot::send_text_message(
@@ -69,4 +79,9 @@ Error: {error:?}",
         }
         Err(error) => Error(format!("{error:?}")),
     }
+}
+
+#[derive(Serialize)]
+struct BtcDepositEventPayload {
+    amount: u64,
 }