+++ frontend/app/src/components/home/communities/details/ChannelCard.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+    import DotsVertical from "svelte-material-icons/DotsVertical.svelte";
+    import DeleteOutline from "svelte-material-icons/DeleteOutline.svelte";
     import Avatar from "../../../Avatar.svelte";
     import AccountMultiple from "svelte-material-icons/AccountMultiple.svelte";
     import {
@@ -15,13 +17,28 @@
     import AccessGateIcon from "../../access/AccessGateIcon.svelte";
     import { popRightPanelHistory } from "../../../../stores/rightPanel";
     import Markdown from "../../Markdown.svelte";
+    import { iconSize } from "@src/stores/iconSize";
+    import MenuIcon from "@src/components/MenuIcon.svelte";
+    import Menu from "@src/components/Menu.svelte";
+    import { i18nKey } from "@src/i18n/i18n";
+    import Translatable from "@src/components/Translatable.svelte";
+    import MenuItem from "@src/components/MenuItem.svelte";
+    import HoverIcon from "@src/components/HoverIcon.svelte";
 
-    export let channel: ChannelMatch;
+    interface Props {
+        channel: ChannelMatch;
+        onDeleteChannel: () => void;
+    }
+
+    let { channel, onDeleteChannel }: Props = $props();
 
     const client = getContext<OpenChat>("client");
 
+    let canDeleteChannel = $derived(client.canDeleteChannel(channel.id));
+
     function selectChannel(match: ChannelMatch) {
         if ($selectedCommunity === undefined) return;
+        if (!match.public) return;
         if ($mobileWidth) {
             popRightPanelHistory();
         }
@@ -29,16 +46,23 @@
     }
 </script>
 
-<div class="details" on:click={() => selectChannel(channel)}>
+<!-- svelte-ignore a11y_click_events_have_key_events -->
+<!-- svelte-ignore a11y_no_static_element_interactions -->
+<div class:clickable={channel.public} class="details" onclick={() => selectChannel(channel)}>
     <div class="avatar">
         <Avatar
             url={client.groupAvatarUrl({ id: channel.id, ...channel.avatar }, $selectedCommunity)}
             size={AvatarSize.Default} />
     </div>
     <div class="channel-text">
-        <h3 class="channel-name">
-            {channel.name}
-        </h3>
+        <div class="channel-name">
+            {#if !channel.public}
+                <div class="private"></div>
+            {/if}
+            <h3>
+                {channel.name}
+            </h3>
+        </div>
         {#if channel.description !== ""}
             <div class="desc">
                 <Markdown text={channel.description} oneLine suppressLinks />
@@ -64,6 +88,30 @@
             </div>
         </div>
     </div>
+    {#if canDeleteChannel}
+        <div class="menu">
+            <MenuIcon position={"bottom"} align={"end"}>
+                {#snippet menuIcon()}
+                    <HoverIcon>
+                        <DotsVertical size={$iconSize} color={"var(--icon-inverted-txt)"} />
+                    </HoverIcon>
+                {/snippet}
+                {#snippet menuItems()}
+                    <Menu>
+                        <MenuItem warning onclick={onDeleteChannel}>
+                            {#snippet icon()}
+                                <DeleteOutline size={$iconSize} color={"var(--menu-warn)"} />
+                            {/snippet}
+                            {#snippet text()}
+                                <Translatable
+                                    resourceKey={i18nKey("deleteGroup", undefined, "channel")} />
+                            {/snippet}
+                        </MenuItem>
+                    </Menu>
+                {/snippet}
+            </MenuIcon>
+        </div>
+    {/if}
 </div>
 
 <style lang="scss">
@@ -75,16 +123,18 @@
         transition:
             background-color ease-in-out 100ms,
             border-color ease-in-out 100ms;
-        cursor: pointer;
+
+        &.clickable {
+            cursor: pointer;
+            @media (hover: hover) {
+                &:hover {
+                    background-color: var(--chatSummary-hv);
+                }
+            }
+        }
 
         @include mobile() {
             padding: 0 toRem(10);
-        }
-
-        @media (hover: hover) {
-            &:hover {
-                background-color: var(--chatSummary-hv);
-            }
         }
 
         .desc {
@@ -122,9 +172,28 @@
             width: 100%;
         }
 
+        .channel-name {
+            display: flex;
+            align-items: center;
+            gap: $sp2;
+            @include ellipsis();
+            h3 {
+                @include font(bold, normal, fs-100);
+            }
+        }
+
         .channel-name,
         .desc {
             margin-bottom: $sp2;
         }
+    }
+
+    .private {
+        background-repeat: no-repeat;
+        $size: 12px;
+        flex: 0 0 $size;
+        width: $size;
+        height: $size;
+        background-image: url("/assets/locked.svg");
     }
 </style>